requirements.txt
# ── Core Orchestration ──────────────────────────────────────────────
langgraph==0.2.28
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-core==0.3.15

# ── LLM & Structured Outputs ────────────────────────────────────────
anthropic==0.40.0
pydantic==2.9.2
pydantic-settings==2.6.1

# ── GitHub Integration ──────────────────────────────────────────────
PyGithub==2.5.0

# ── Web Framework ───────────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# ── Docker SDK (for running sandbox containers) ──────────────────────
docker==7.1.0

# ── Utilities ───────────────────────────────────────────────────────
httpx==0.27.2
tenacity==9.0.0       # retry logic with backoff
structlog==24.4.0     # structured JSON logging
rich==13.9.4          # pretty terminal output during dev